requirements.txt
Django==5.2.8
PyMySQL==1.1.1
Pillow==10.4.0
python-dotenv==1.0.1
openai==1.47.0
reportlab==4.2.5
requests==2.32.3
django-allauth==64.2.1
